Exception Handling Done

+++ NoBS/src/main/java/com/example/demo/Product/commandhandlers/DeleteProductCommandHandler.java
@@ -1,11 +1,14 @@
 package com.example.demo.Product.commandhandlers;
 
 import com.example.demo.Command;
+import com.example.demo.Exceptions.ProductNotFoundException;
 import com.example.demo.Product.Model.Product;
 import com.example.demo.Product.ProductRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
+
+import java.util.Optional;
 
 @Service
 public class DeleteProductCommandHandler implements Command<Integer, ResponseEntity> {
@@ -13,6 +16,10 @@ public class DeleteProductCommandHandler implements Command<Integer, ResponseEnt
     private ProductRepository productRepository;
     @Override
     public ResponseEntity<ResponseEntity> execute(Integer id) {
+        Optional<Product> optionalProduct = productRepository.findById(id);
+        if(optionalProduct.isEmpty()){
+            throw new ProductNotFoundException();
+        }
         Product product = productRepository.findById(id).get();
         productRepository.delete(product);
         return ResponseEntity.ok().build();
